ISO: move mDocZk code out of wallet package
