+++ vck/src/commonMain/kotlin/at/asitplus/iso/zk/IsoMdocZkProofFactory.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.isoMdocZk
+package at.asitplus.iso.zk
 
 import at.asitplus.KmmResult
 import at.asitplus.iso.SessionTranscript
